==> research/2_embedding_files_creation.ipynb <==
%cd /
from google.colab import drive
drive.mount('/content/gdrive/')

!ln -s /content/gdrive/My\ Drive/ ./mydrive

%cd /content/gdrive/MyDrive/projekt_inzynierski

!pip install sentence-transformers -q

import pandas as pd
from sentence_transformers import SentenceTransformer
from tqdm import tqdm
import torch

df = pd.read_csv('df_kaggle_enriched_cleaned.csv')
df.sample(5)

df['tags'] = df['tags'].str.replace('_', ' ')
df['tags'] = df['tags'].fillna("")
df.sample(5)

device = 'cuda' if torch.cuda.is_available() else 'cpu'

model = SentenceTransformer('intfloat/multilingual-e5-base', device=device)

required_cols = ['track_id', 'name', 'artist', 'tags']

tags = df['tags']
tags

embeddings = model.encode(
    tags.tolist(),
    batch_size=64,
    convert_to_numpy=True,
    normalize_embeddings=True,
    show_progress_bar=True
)

embeddings

df_embeddings = df[['track_id', 'name', 'artist', 'tags']].copy()
df_embeddings['tags_embedding'] = embeddings.tolist()
df_embeddings

import numpy as np

np.save('./tag_embeddings.npy', embeddings)

df_embeddings.to_parquet('./df_embeddings.parquet', index=False)

df['tags_list'] = df['tags'].apply(lambda x: [tag.strip().lower() for tag in x.split(',')])

all_tags = [tag for tags in df['tags_list'] for tag in tags]
unique_tags = list(set(all_tags))

unique_tags = [tag for tag in unique_tags if tag]

tag_texts = [f"passage: {t}" for t in unique_tags]

tag_embeddings = model.encode(
    tag_texts,
    batch_size=64,
    convert_to_numpy=True,
    normalize_embeddings=True,
    show_progress_bar=True
).astype("float32")

df_tag_embeddings = pd.DataFrame({
    'tag': unique_tags,
    'tag_embedding': tag_embeddings.tolist()
})

np.save('./unique_tag_embeddings.npy', tag_embeddings)
df_tag_embeddings.to_parquet('./df_unique_tag_embeddings.parquet', index=False)
